requirements.txt
# FastAPI y servidor
fastapi==0.117.1
uvicorn==0.37.0
python-multipart==0.0.20
python-dotenv==1.1.1
Jinja2==3.1.6
starlette==0.48.0

# Base de datos
SQLAlchemy==2.0.43
psycopg2-binary==2.9.10

# Autenticación JWT
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
bcrypt==4.2.1

# Sesiones
itsdangerous==2.2.0

# Validación
pydantic==2.11.9
email-validator==2.3.0

# Utilidades
certifi==2025.8.3

# Supabase (opcional)
supabase==2.9.0

# Procesamiento de imágenes
Pillow==11.0.0

